LPS-60606 Move ExportImportPortletConfigurationIcon classes to ExportImportWeb and removed addon since it's no longer used
