--- modules/addons/portlet-configuration-icon/portlet-configuration-icon-export-import/bnd.bnd
@@ -1,4 +1,0 @@
-Bundle-Name: Liferay Portlet Configuration Icon Export Import
-Bundle-SymbolicName: com.liferay.portlet.configuration.icon.exportimport
-Bundle-Version: 1.0.0
-Include-Resource: classes
