+++ modules/apps/export-import/export-import-web/src/main/java/com/liferay/exportimport/web/configuration/icon/ExportImportPortletConfigurationIconFactory.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.configuration.icon.exportimport;
+package com.liferay.exportimport.web.configuration.icon;
 
 import com.liferay.portal.kernel.portlet.configuration.icon.BasePortletConfigurationIconFactory;
 import com.liferay.portal.kernel.portlet.configuration.icon.PortletConfigurationIcon;
